+++ admin/users.php
@@ -135,7 +135,8 @@ function get_sort_link($column, $current_sort, $current_order) {
             <a href="media_logs.php">Log Media</a> |
             <a href="channels.php">Channel</a> |
             <a href="database.php">Database</a> |
-            <a href="logs.php">Logs</a>
+            <a href="logs.php">Logs</a> |
+            <a href="telegram_logs.php">Log Error Telegram</a>
         </nav>
 
         <h1>Manajemen Pengguna</h1>